+++ src/main/java/blackjack/BlackJackApplication.java
@@ -1,0 +1,11 @@
+package blackjack;
+
+import blackjack.controller.BlackJackController;
+
+public class BlackJackApplication {
+
+    public static void main(final String... args) {
+        BlackJackController blackJackController = new BlackJackController();
+        blackJackController.run();
+    }
+}
